G1106_S12

+++ G1106_S05/app/src/main/java/ro/ase/ie/g1106_s05/activities/MainActivity.java
@@ -137,7 +137,7 @@ public class MainActivity extends AppCompatActivity implements IMovieEvents {
             Map<Integer, ArrayList<Movie>> movieOptions = movieAdapter.getMovieOptions();
             ArrayList<Movie> movies = movieOptions.get(R.id.rbPersist);
             for (Movie movie : movies) {
-                int id = movieDao.getMovieByTitleAndRelease(movie.getTitle(), movie.getRelease().getTime());
+                long id = movieDao.getMovieByTitleAndRelease(movie.getTitle(), movie.getRelease().getTime());
                 movie.setMovieId(id == 0 ? null : id);
                 movieDao.insert(movie);
             }
